requirements.txt
opencv-python
pillow
pdf2image
pytesseract
langdetect
pymupdf
